Do not bundle vmlinux.h in the SRPM
