+++ backends/ebpf/stub.go
@@ -29,3 +29,13 @@ func (b *EbpfBackend) Run(<-chan struct{}, <-chan glowdTypes.FlowID) {
 func (b *EbpfBackend) Cleanup() error {
 	return nil
 }
+
+// This type definition is needed so that the cmd utils won't
+// complain. Should we maybe relocate this definition someplace
+// else?
+type FlowFourTuple struct {
+	IPv6Hi  uint64
+	IPv6Lo  uint64
+	DstPort uint16
+	SrcPort uint16
+}
