+++ cmd/subcmd/ebpf.go
@@ -1,3 +1,5 @@
+//go:build linux && cgo
+
 package subcmd
 
 import (
