+++ enrichment/skops/cgroup_test.go
@@ -1,3 +1,5 @@
+//go:build linux && cgo
+
 package skops
 
 import (
