+++ enrichment/skops/ebpf.go
@@ -1,3 +1,5 @@
+//go:build linux && cgo
+
 package skops
 
 import (
