+++ enrichment/skops/stub.go
@@ -1,0 +1,3 @@
+//go:build darwin || !cgo
+
+package skops
